+++ R/geolayer.R
@@ -94,7 +94,7 @@ as_geolayer.star_database <- function(db,
   # data$value <- as.numeric(data$value)
   data <- data |>
     tidyr::spread("variable", "value")
-  data <- dplyr::left_join(data, geo, by = gt_att)
+  data <- dplyr::full_join(data, geo, by = gt_att)
   data <- sf::st_as_sf(data)
 
   structure(list(
@@ -106,11 +106,14 @@ as_geolayer.star_database <- function(db,
 }
 
 
-#' Get geographic layer
+#' Get geographic information layer
 #'
-#' Get the geographic layer from a `geolayer` object.
+#' Get the geographic information layer from a `geolayer` object.
+#'
+#' By default, rows that are NA for all variables are eliminated.
 #'
 #' @param gl A `geolayer` object.
+#' @param keep_all_variables_na A boolean, keep rows with all variables NA.
 #'
 #' @return A `sf` object.
 #'
@@ -122,15 +125,20 @@ as_geolayer.star_database <- function(db,
 #'   as_geolayer()
 #'
 #' l <- gl |>
-#'   get_geolayer()
+#'   get_layer()
 #'
 #' @export
-get_geolayer <- function(gl)
-  UseMethod("get_geolayer")
+get_layer <- function(gl, keep_all_variables_na)
+  UseMethod("get_layer")
 
-#' @rdname get_geolayer
+#' @rdname get_layer
 #' @export
-get_geolayer.geolayer <- function(gl) {
+get_layer.geolayer <- function(gl, keep_all_variables_na = FALSE) {
+  variable <- unique(gl$variables$variable)
+  if (!keep_all_variables_na) {
+    gl$geolayer <- gl$geolayer |>
+      dplyr::filter(!dplyr::if_all(tidyselect::all_of(variable), is.na))
+  }
   gl$geolayer
 }
 
@@ -212,13 +220,60 @@ set_variables.geolayer <- function(gl, variables, keep_all_variables_na = FALSE)
   vars <- intersect(names(gl$geolayer), c(gl$geoattribute, variable))
   gl$geolayer <- gl$geolayer |>
     dplyr::select(tidyselect::all_of(vars))
-  if (!keep_all_variables_na) {
-    gl$geolayer <- gl$geolayer |>
-      dplyr::filter(!dplyr::if_all(tidyselect::all_of(variable), is.na))
-  }
+  gl$geolayer <- gl |>
+    get_layer(keep_all_variables_na)
   gl
 }
 
+
+#' Get variable description
+#'
+#' Obtain a description of the variables whose name is indicated. If no name is
+#' indicated, all are returned.
+#'
+#' Using the parameter `only_values`, we can obtain only the combination of values
+#' or also the combination of names with values.
+#'
+#' @param gl A `geolayer` object.
+#' @param name A string vector.
+#' @param only_values A boolean, add names to component values.
+#'
+#' @return A string vector.
+#'
+#' @family query functions
+#'
+#' @examples
+#'
+#' gl <- mrs_db_geo |>
+#'   as_geolayer()
+#'
+#' vd <- gl |>
+#'   get_variable_description()
+#'
+#' @export
+get_variable_description <- function(gl, name, only_values)
+  UseMethod("get_variable_description")
+
+#' @rdname get_variable_description
+#' @export
+get_variable_description.geolayer <- function(gl, name = NULL, only_values = FALSE) {
+  if (is.null(name)) {
+    name <- gl$variables$variable
+  } else {
+    name <- name[name %in% gl$variables$variable]
+  }
+  vars <- gl$variables[gl$variables$variable %in% name, -1]
+  vars <- as.data.frame(vars)
+  if (!only_values) {
+    col_names <- colnames(vars)
+    for (j in seq_along(col_names)) {
+      vars[, j] <- paste0(col_names[j], " = ", vars[, j])
+    }
+  }
+  des <- apply(vars, 1, paste, collapse=", ")
+  names(des) <- name
+  des
+}
 
 
 #' Save as `GeoPackage`
@@ -228,6 +283,8 @@ set_variables.geolayer <- function(gl, variables, keep_all_variables_na = FALSE)
 #'
 #' If the file name is not indicated, it defaults to the name of the geovariable.
 #'
+#' By default, rows that are NA for all variables are eliminated.
+#'
 #' The `GeoPackage` format only allows defining a maximum of 1998 columns. If the
 #' number of variables and columns in the geographic layer exceeds this number,
 #' it cannot be saved in this format.
@@ -235,6 +292,7 @@ set_variables.geolayer <- function(gl, variables, keep_all_variables_na = FALSE)
 #' @param gl A `geolayer` object.
 #' @param dir A string.
 #' @param name A string, file name.
+#' @param keep_all_variables_na A boolean, keep rows with all variables NA.
 #'
 #' @return A string, file name.
 #'
@@ -249,12 +307,12 @@ set_variables.geolayer <- function(gl, variables, keep_all_variables_na = FALSE)
 #'   as_GeoPackage(dir = tempdir())
 #'
 #' @export
-as_GeoPackage <- function(gl, dir, name)
+as_GeoPackage <- function(gl, dir, name, keep_all_variables_na)
   UseMethod("as_GeoPackage")
 
 #' @rdname as_GeoPackage
 #' @export
-as_GeoPackage.geolayer <- function(gl, dir = NULL, name = NULL) {
+as_GeoPackage.geolayer <- function(gl, dir = NULL, name = NULL, keep_all_variables_na = FALSE) {
   stopifnot(
     "The maximum number of columns supported by this format (1998 cols.) has been exceeded." = ncol(gl$geolayer) < 1999
   )
@@ -267,8 +325,11 @@ as_GeoPackage.geolayer <- function(gl, dir = NULL, name = NULL) {
   name <- tools::file_path_sans_ext(name)
   file <- paste0(dir, name, '.gpkg')
 
+  layer <- gl |>
+    get_layer(keep_all_variables_na)
+
   sf::st_write(
-    obj = gl$geolayer,
+    obj = layer,
     dsn = file,
     layer = "geolayer",
     append = FALSE,
@@ -306,6 +367,15 @@ same_granularity_facts <- function(db, names) {
   TRUE
 }
 
+#' Unify facts and dimensions in a flat table
+#'
+#' @param db A `star_database` object.
+#' @param dimension A vector of strings, dimension names.
+#' @param include_nrow_agg A boolean.
+#'
+#' @return A `tibble`.
+#'
+#' @keywords internal
 unify_facts_and_dimensions <- function(db, dimension, include_nrow_agg) {
   facts <- NULL
   for (f in names(db$facts)) {